+++ apps/core-app/src/App.tsx
@@ -1,0 +1,14 @@
+import { DaoHausNav } from '@daohaus/daohaus-connect-feature';
+import { H1 } from '@daohaus/ui';
+import { OuterLayout } from '@daohaus/ui';
+
+export function App() {
+  return (
+    <OuterLayout>
+      <DaoHausNav />
+      <H1>Header!</H1>
+    </OuterLayout>
+  );
+}
+
+export default App;
