+++ apps/core-app/src/main.tsx
@@ -4,14 +4,14 @@ import { StrictMode } from 'react';
 import * as ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 
-import App from './app/App';
+import App from './App';
 
 ReactDOM.render(
   <StrictMode>
     <BrowserRouter>
       <HausThemeProvider>
         <HausConnectProvider>
-          <App title="fuck" />
+          <App />
         </HausConnectProvider>
       </HausThemeProvider>
     </BrowserRouter>
