+++ apps/summon-app/src/layouts/FormLayouts.tsx
@@ -1,87 +1,7 @@
-import { Divider, H5, ParMd, widthQuery } from '@daohaus/ui';
+import { widthQuery } from '@daohaus/ui';
 import styled from 'styled-components';
 
-type SegmentType = {
-  title?: string | React.ReactNode;
-  description?: string | React.ReactNode;
-  formArea: React.ReactNode;
-};
-
-const StyledFormSegment = styled.section`
-  width: 100%;
-  margin-bottom: 2.4rem;
-  .segment-title {
-    margin-bottom: 3.2rem;
-  }
-  .segment-description {
-    margin-bottom: 3rem;
-  }
-`;
-
-export const FormSegment = ({ title, description, formArea }: SegmentType) => {
-  return (
-    <StyledFormSegment>
-      <H5 className="segment-title">{title}</H5>
-      <ParMd className="segment-description">{description}</ParMd>
-      <div>{formArea}</div>
-      <Divider />
-    </StyledFormSegment>
-  );
-};
-
-const StyledSplitColumn = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  flex-direction: column;
-`;
-type Row = {
-  rowID: string;
-  left: React.ReactNode;
-  right: React.ReactNode;
-};
-type SplitColumnProps = {
-  singleRow?: Row;
-  rows?: Row[];
-};
-
-export const SplitColumn = ({ rows, singleRow }: SplitColumnProps) => {
-  return (
-    <StyledSplitColumn>
-      {singleRow && <Row {...singleRow} />}
-      {rows?.map((row) => {
-        return <Row key={row.rowID} {...row} />;
-      })}
-    </StyledSplitColumn>
-  );
-};
-
-const StyledRow = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  margin-bottom: 3.6rem;
-  .left-section,
-  .right-section {
-    width: 24rem;
-    @media ${widthQuery.sm} {
-      width: 100%;
-      margin-bottom: 1.6rem;
-    }
-  }
-  @media ${widthQuery.sm} {
-    flex-direction: column;
-  }
-`;
-
-const Row = ({ left, right }: Row) => {
-  return (
-    <StyledRow>
-      <div className="left-section">{left}</div>
-      <div className="right-section">{right}</div>
-    </StyledRow>
-  );
-};
+// LAYOUT REVIEW There's more to extract here. We need to see how the sub-layouts work first.
 
 export const TextAreaSection = styled.div`
   width: 100%;
