+++ apps/summon-app/src/layouts/TimingSegment.tsx
@@ -1,5 +1,5 @@
+import { FormSegment, SplitColumn } from '@daohaus/ui';
 import { TimePicker } from '../components/TimePicker/TimePicker';
-import { FormSegment, SplitColumn } from '../layouts/FormLayouts';
 import { INFO_COPY } from '../utils/content';
 import { FORM_KEYS } from '../utils/formKeys';
