+++ libs/ui/.storybook/preview.js
@@ -78,4 +78,5 @@ const WithFormProvider = (Story, context) => {
     </FormProvider>
   );
 };
+
 export const decorators = [withThemeProvider, WithFormProvider];
