+++ libs/ui/src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './atoms';
 export * from './molecules';
 export * from './organisms';
+export * from './layouts';
