+++ libs/ui/src/components/layouts/FormSegment/FormSegment.styles.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const StyledFormSegment = styled.section`
+  width: 100%;
+  margin-bottom: 2.4rem;
+
+  .segment-title {
+    margin-bottom: 3.2rem;
+  }
+
+  .segment-description {
+    margin-bottom: 3rem;
+  }
+`;
