+++ libs/ui/src/components/layouts/OuterLayout/index.ts
@@ -1,0 +1,1 @@
+export * from './OuterLayout';
