+++ libs/ui/src/components/layouts/SplitColumn/SplitColumn.tsx
@@ -1,0 +1,34 @@
+import { StyledRow, StyledSplitColumn } from './SplitColumn.styles';
+
+// REVIEW: No stories until there's a solid form sub-layout
+
+type SplitColumnProps = {
+  rows: Row[] | Row;
+};
+export const SplitColumn = ({ rows }: SplitColumnProps) => {
+  return (
+    <StyledSplitColumn>
+      {Array.isArray(rows) ? (
+        rows.map((row) => {
+          return <Row key={row.rowID} {...row} />;
+        })
+      ) : (
+        <Row {...rows} />
+      )}
+    </StyledSplitColumn>
+  );
+};
+
+type Row = {
+  rowID: string;
+  left: React.ReactNode;
+  right: React.ReactNode;
+};
+const Row = ({ left, right }: Row) => {
+  return (
+    <StyledRow>
+      <div className="left-section">{left}</div>
+      <div className="right-section">{right}</div>
+    </StyledRow>
+  );
+};
